handle components UI and logic
