+++ react-client/src/components/FlowBuilder/componentsPopover.js
@@ -25,7 +25,9 @@ class ComponentsPopover extends React.Component {
               whiteSpace: 'normal',
               width: '235px'
             }}
-            type="button" className="btn btn-outline-brand"
+            onClick={() => this.props.showAddComponentModal('text')}
+            type="button"
+            className="btn btn-outline-brand"
           >
               + Add Text Component
           </button>
@@ -36,6 +38,7 @@ class ComponentsPopover extends React.Component {
               whiteSpace: 'normal',
               width: '235px'
             }}
+            onClick={() => this.props.showAddComponentModal('media')}
             type="button"
             className="btn btn-outline-brand"
           >
@@ -48,6 +51,7 @@ class ComponentsPopover extends React.Component {
               whiteSpace: 'normal',
               width: '235px'
             }}
+            onClick={() => this.props.showAddComponentModal('card')}
             type="button"
             className="btn btn-outline-brand"
           >
@@ -60,6 +64,7 @@ class ComponentsPopover extends React.Component {
               whiteSpace: 'normal',
               width: '235px'
             }}
+            onClick={() => this.props.showAddComponentModal('audio')}
             type="button"
             className="btn btn-outline-brand"
           >
@@ -72,6 +77,7 @@ class ComponentsPopover extends React.Component {
               whiteSpace: 'normal',
               width: '235px'
             }}
+            onClick={() => this.props.showAddComponentModal('file')}
             type="button"
             className="btn btn-outline-brand"
           >
@@ -84,6 +90,7 @@ class ComponentsPopover extends React.Component {
               whiteSpace: 'normal',
               width: '235px'
             }}
+            onClick={() => this.props.showAddComponentModal('video')}
             type="button"
             className="btn btn-outline-brand"
           >
@@ -96,6 +103,7 @@ class ComponentsPopover extends React.Component {
               whiteSpace: 'normal',
               width: '235px'
             }}
+            onClick={() => this.props.showAddComponentModal('link')}
             type="button"
             className="btn btn-outline-brand"
           >
@@ -110,7 +118,8 @@ class ComponentsPopover extends React.Component {
 ComponentsPopover.propTypes = {
   'showPopover': PropTypes.bool.isRequired,
   'togglePopover': PropTypes.func.isRequired,
-  'targetId': PropTypes.string.isRequired
+  'targetId': PropTypes.string.isRequired,
+  'showAddComponentModal': PropTypes.func.isRequired
 }
 
 export default ComponentsPopover
