+++ react-client/src/components/FlowBuilder/startingStep.js
@@ -1,11 +1,24 @@
 import React from "react"
-// import PropTypes from 'prop-types'
+import PropTypes from 'prop-types'
 import COMPONENTSAREA from './componentsArea'
 
 class StartingStep extends React.Component {
   constructor (props, context) {
     super(props, context)
     this.state = {}
+    this.getItems = this.getItems.bind(this)
+  }
+
+  getItems () {
+    let items = []
+    for (let i = 0; i < this.props.linkedMessages[0].messageContent.length; i++) {
+      let component = this.props.getComponent(this.props.linkedMessages[0].messageContent[i]).component
+      items.push({content: component})
+    }
+    if (items.length > 0) {
+      items.push(this.props.getQuickReplies())
+    }
+    return items
   }
 
   render () {
@@ -19,6 +32,8 @@ class StartingStep extends React.Component {
         <div style={{border: '1px solid #34bfa3'}} className='card-body'>
           <COMPONENTSAREA
             targetId='starting-step-add-component'
+            showAddComponentModal={this.props.showAddComponentModal}
+            getItems={this.getItems}
           />
         </div>
       </div>
@@ -27,7 +42,11 @@ class StartingStep extends React.Component {
 }
 
 StartingStep.propTypes = {
-
+  'showAddComponentModal': PropTypes.func.isRequired,
+  'getQuickReplies': PropTypes.func.isRequired,
+  'getComponent': PropTypes.func.isRequired,
+  'linkedMessages': PropTypes.array.isRequired,
+  'unlinkedMessages': PropTypes.array.isRequired
 }
 
 export default StartingStep
